SphericalHarmonics.jl

SphericalHarmonics.NorthPole — Type

SphericalHarmonics.NorthPole <: SphericalHarmonics.Pole

The angle $\theta = 0$ in spherical polar coordinates.

SphericalHarmonics.Pole — Type

SphericalHarmonics.Pole <: Real

Supertype of NorthPole and SouthPole.

SphericalHarmonics.SouthPole — Type

SphericalHarmonics.SouthPole <: SphericalHarmonics.Pole

The angle $\theta = π$ in spherical polar coordinates.

SphericalHarmonics.allocate_p — Method

SphericalHarmonics.allocate_p([T::Type = Float64], L::Integer)

Allocate an array large enough to store an entire set of Associated Legendre Polynomials $\bar{P}_l^m(x)$ of maximum degree $L$.

SphericalHarmonics.allocate_y — Method

SphericalHarmonics.allocate_y([T::Type = ComplexF64], L::Integer)

Allocate an array large enough to store an entire set of spherical harmonics $Y_{l,m}(θ,φ)$ of maximum degree $L$.

SphericalHarmonics.computePlmcostheta! — Method

computePlmcostheta!(P::AbstractVector{<:Real}, θ::Real, lmax::Integer, coeff::AbstractMatrix)
computePlmcostheta!(P::AbstractVector{<:Real}, θ::SphericalHarmonics.Pole, lmax::Integer)

Compute an entire set of normalized Associated Legendre Polynomials $\bar{P}_l^m(cos(θ))$ using the given coefficients, and store in the array P. The matrix coeff may be computed using compute_coefficients.

SphericalHarmonics.computePlmcostheta — Method

computePlmcostheta(θ::Real; lmax::Integer)

Compute an entire set of normalized Associated Legendre Polynomials $\bar{P}_l^m(cos(θ))$ where $0 ≤ l ≤ l_\mathrm{max}$ and $0 ≤ m ≤ l$ for colatitude $\theta$.

The polynomials are normalized as

\[\bar{P}_{\ell m} = \sqrt{\frac{(2\ell + 1)(\ell-m)!}{2\pi (\ell+m)!}} P_{\ell m}.\]

Returns an SHVector that may be indexed using (l,m) pairs aside from the canonical indexing with Ints.

Examples

julia> P = computePlmcostheta(pi/2, 1)
3-element SHArray(::Array{Float64,1}, (ML(0:1, 0:1),)):
  0.3989422804014327
  4.231083042742082e-17
 -0.4886025119029199

julia> P[(0,0)]
0.3989422804014327

SphericalHarmonics.computePlmx! — Method

computePlmx!(P::AbstractVector{<:Real}, x::Real, lmax::Integer, coeff::AbstractMatrix)

Compute an entire set of normalized Associated Legendre Polynomials $\bar{P}_l^m(x)$ using the given coefficients, and store in the array P. The matrix coeff may be computed using compute_coefficients.

The argument x needs to lie in $-1 ≤ x ≤ 1$. The function implicitly assumes that $x = \cos(\theta)$ where $0 ≤ \theta ≤ π$.

SphericalHarmonics.computePlmx — Method

computePlmx(x::Real; lmax::Integer)

Compute an entire set of normalized Associated Legendre Polynomials $\bar{P}_l^m(x)$ where $0 ≤ l ≤ l_\mathrm{max}$ and $0 ≤ m ≤ l$.

The argument x needs to lie in $-1 ≤ x ≤ 1$. The function implicitly assumes that $x = \cos(\theta)$ where $0 ≤ \theta ≤ π$.

SphericalHarmonics.computeYlm! — Method

computeYlm!(Y::AbstractVector{<:Complex}, θ::Real, φ::Real; lmax::Integer, [SHtype = ComplexHarmonics()])

Compute an entire set of spherical harmonics $Y_{l,m}(θ,φ)$ for $0 ≤ l ≤ l_\mathrm{max}$, and store them in the array Y.

The optional argument SHtype may be used to choose between real and complex harmonics. To compute real spherical harmonics, set this to RealHarmonics().

SphericalHarmonics.computeYlm! — Method

computeYlm!(Y::AbstractVector, P::AbstractVector{<:Real}, θ::Real, φ::Real, L::Integer, [SHtype = ComplexHarmonics()])

Compute an entire set of spherical harmonics $Y_{l,m}(θ,φ)$ using the precomputed associated Legendre Polynomials $\bar{P}_l^m(x = \cos(θ))$, and store in the array Y. The array P may be computed using computePlmcostheta.

The optional argument SHtype may be used to choose between real and complex harmonics. To compute real spherical harmonics, set this to RealHarmonics().

SphericalHarmonics.computeYlm — Method

computeYlm(θ::Real, ϕ::Real; lmax::Integer, [SHtype = ComplexHarmonics()])

Compute an entire set of spherical harmonics $Y_{l,m}(θ,ϕ)$ for $0 ≤ l ≤ l_\mathrm{max}$ and $-l ≤ m ≤ l$, for colatitude $\theta$ and azimuth $\phi$.

Returns an SHVector that may be indexed using (l,m) pairs aside from the canonical indexing with Ints.

The optional argument SHtype may be used to choose between real and complex harmonics. To compute real spherical harmonics, set this to RealHarmonics().

Examples

julia> Y = computeYlm(pi/2, 0, 1)
4-element SHArray(::Array{Complex{Float64},1}, (ML(0:1, -1:1),)):
     0.2820947917738782 + 0.0im
     0.3454941494713355 - 0.0im
 2.9918275112863375e-17 + 0.0im
    -0.3454941494713355 - 0.0im

julia> Y[(1,-1)]
0.3454941494713355 - 0.0im

julia> Y = computeYlm(big(pi)/2, 0, 1) # Arbitrary precision
4-element SHArray(::Array{Complex{BigFloat},1}, (ML(0:1, -1:1),)):
    0.2820947917738781434740397257803862929220253146644994284220428608553212342207478 + 0.0im
    0.3454941494713354800004725866746756805800203549224377345215348347601343937267067 - 0.0im
 2.679783085063171668225419916118067917387251852939708540164955895366691604430101e-78 + 0.0im
   -0.3454941494713354800004725866746756805800203549224377345215348347601343937267067 - 0.0im

julia> computeYlm(SphericalHarmonics.NorthPole(), 0, 1)
4-element SHArray(::Array{Complex{Float64},1}, (ML(0:1, -1:1),)):
 0.2820947917738782 + 0.0im
                0.0 + 0.0im
   0.48860251190292 + 0.0im
                0.0 + 0.0im

SphericalHarmonics.compute_coefficients — Method

SphericalHarmonics.compute_coefficients(L::Integer)

Precompute coefficients $a_l^m$ and $b_l^m$ for all $l ≤ L$ and $0 ≤ m ≤ l$.