SphericalHarmonics.jl

SphericalHarmonics.NorthPole — Type

SphericalHarmonics.NorthPole <: SphericalHarmonics.Pole

The angle $\theta = 0$ in spherical polar coordinates.

SphericalHarmonics.Pole — Type

SphericalHarmonics.Pole <: Real

Supertype of NorthPole and SouthPole.

SphericalHarmonics.SouthPole — Type

SphericalHarmonics.SouthPole <: SphericalHarmonics.Pole

The angle $\theta = π$ in spherical polar coordinates.

SphericalHarmonics.allocate_p — Method

SphericalHarmonics.allocate_p([T::Type = Float64], L::Integer)

Allocate an array large enough to store an entire set of Associated Legendre Polynomials $\bar{P}_l^m(x)$ of maximum degree $L$.

SphericalHarmonics.allocate_y — Function

SphericalHarmonics.allocate_y([T::Type = ComplexF64], L::Integer)

Allocate an array large enough to store an entire set of spherical harmonics $Y_{l,m}(θ,ϕ)$ of maximum degree $L$.

SphericalHarmonics.computePlmcostheta! — Method

computePlmcostheta!(P::AbstractVector{<:Real}, θ::Real, lmax::Integer, coeff::AbstractMatrix)
computePlmcostheta!(P::AbstractVector{<:Real}, θ::SphericalHarmonics.Pole, lmax::Integer)

Compute an entire set of normalized Associated Legendre Polynomials $\bar{P}_l^m(\cos θ)$ using the given coefficients, and store in the array P. The matrix coeff may be computed using compute_coefficients.

See computePlmcostheta for the normalization used.

computePlmcostheta!(P::AbstractVector{<:Real}, θ::Real, lmax::Integer, m::Integer, coeff::AbstractMatrix)

Compute the Associated Legendre Polynomials $\bar{P}_l^m(\cos θ)$ for for the specified $m$ and all $l$ lying in $|m| ≤ l ≤ l_\mathrm{max}$. The array P needs to be large enough to hold all the polynomials for $0 ≤ l ≤ l_\mathrm{max}$ and $0 ≤ m ≤ l$, as the recursive evaluation requires the computation of extra elements. Pre-existing values in P may be overwritten, even for azimuthal orders not equal to $m$.

SphericalHarmonics.computePlmcostheta — Method

computePlmcostheta(θ::Real; lmax::Integer, [m::Integer])
computePlmcostheta(θ::Real, lmax::Integer, [m::Integer])

Compute an entire set of normalized Associated Legendre Polynomials $\bar{P}_l^m(\cos θ)$ where $0 ≤ l ≤ l_\mathrm{max}$ and $0 ≤ m ≤ l$ for colatitude $\theta$. If m is provided then only the polynomials corresponding to the specified m are computed.

The polynomials are normalized as

\[\bar{P}_{\ell}^m = \sqrt{\frac{(2\ell + 1)(\ell-m)!}{2\pi (\ell+m)!}} P_{\ell m},\]

where $P_{\ell m}$ are the standard associated Legendre polynomials, and are defined in terms of Legendre polynomials $P_\ell(x)$ as

\[P_{\ell m}\left(x\right)=\left(1-x^{2}\right)^{m/2}\frac{d^{m}}{dx^{m}}P_{\ell}\left(x\right).\]

The normalized polynomials $\bar{P}_{\ell}^m$ satisfy

\[\int_{0}^{\pi} \sin θ d\theta\,\left| \bar{P}_{\ell}^m(\cos θ) \right|^2 = \frac{1}{\pi}\]

Info

The Condon-Shortley phase factor $(-1)^m$ is not included in the definition of the polynomials.

Returns an SHVector that may be indexed using (l,m) pairs aside from the canonical indexing with Ints.

The precision of the result may be increased by using arbitrary-precision arguments.

Examples

julia> P = computePlmcostheta(pi/2, lmax = 1)
3-element SHArray(::Array{Float64,1}, (ML(0:1, 0:1),)):
  0.3989422804014327
  4.231083042742082e-17
 -0.4886025119029199

julia> P[(0,0)]
0.3989422804014327

julia> P = computePlmcostheta(big(pi)/2, lmax = big(1)) # Arbitrary precision
3-element SHArray(::Array{BigFloat,1}, (ML(0:1, 0:1),)):
  0.3989422804014326779399460599343818684758586311649346576659258296706579258993008
  3.789785583114350800838137317730900078444216599640987847808409161681770236721676e-78
 -0.4886025119029199215863846228383470045758856081942277021382431574458410003616367

SphericalHarmonics.computePlmx! — Method

computePlmx!(P::AbstractVector{<:Real}, x::Real, lmax::Integer, coeff::AbstractMatrix)

Compute an entire set of normalized Associated Legendre Polynomials $\bar{P}_l^m(x)$ using the given coefficients, and store in the array P. The matrix coeff may be computed using compute_coefficients.

The argument x needs to lie in $-1 ≤ x ≤ 1$. The function implicitly assumes that $x = \cos(\theta)$ where $0 ≤ \theta ≤ π$.

See computePlmcostheta for the normalization used.

computePlmx!(P::AbstractVector{<:Real}, x::Real, lmax::Integer, m::Integer, coeff::AbstractMatrix)

Compute the set of normalized Associated Legendre Polynomials $\bar{P}_l^m(x)$ for for the specified $m$ and all $l$ lying in $|m| ≤ l ≤ l_\mathrm{max}$ .

SphericalHarmonics.computePlmx — Method

computePlmx(x::Real; lmax::Integer, [m::Integer])
computePlmx(x::Real, lmax::Integer, [m::Integer])

Compute an entire set of normalized Associated Legendre Polynomials $\bar{P}_l^m(x)$ where $0 ≤ l ≤ l_\mathrm{max}$ and $0 ≤ m ≤ l$. If $m$ is provided then only the polynomials for that azimuthal order are computed.

The argument x needs to lie in $-1 ≤ x ≤ 1$. The function implicitly assumes that $x = \cos(\theta)$ where $0 ≤ \theta ≤ π$.

SphericalHarmonics.computeYlm — Function

computeYlm(θ::Real, ϕ::Real; lmax::Integer, [m::Integer], [m_range = SphericalHarmonics.FullRange], [SHType = SphericalHarmonics.ComplexHarmonics()])
computeYlm(θ::SphericalHarmonics.Pole; lmax::Integer, [m_range = SphericalHarmonics.FullRange], [SHType = SphericalHarmonics.ComplexHarmonics()])

Compute an entire set of spherical harmonics $Y_{l,m}(θ,ϕ)$ for $0 ≤ l ≤ l_\mathrm{max}$ and $-l ≤ m ≤ l$, for colatitude $\theta$ and azimuth $\phi$. If $m$ is provided then only the polynomials for the specified $m$ are computed.

Returns an SHVector that may be indexed using (l,m) pairs aside from the canonical indexing with Ints.

The optional argument m_range decides if the spherical harmonics for negative m values are computed. By default the functions for all values of m are evaluated. Setting m_range to SphericalHarmonics.ZeroTo would result in only functions for m ≥ 0 being evaluated.

The optional argument SHType may be used to choose between real and complex harmonics. To compute real spherical harmonics, set this to SphericalHarmonics.RealHarmonics(). The complex harmonics are defined as

\[Y_{\ell,m}\left(\theta,\phi\right)=\begin{cases} \frac{1}{\sqrt{2}}\bar{P}_{\ell}^{m}\left(\cos\theta\right)\exp\left(im\phi\right), & m\geq0,\\ \left(-1\right)^{m}Y_{\ell,-m}^{*}\left(\theta,\phi\right), & m<0. \end{cases}\]

This definition corresponds to Laplace sphercial harmonics, whereas the quantum mechanical definition prepends a Condon-Shortley phase on the harmonics.

The real spherical harmonics are defined as

\[Y_{\ell,m}\left(\theta,\phi\right)=\begin{cases} \bar{P}_{\ell}^{\left|m\right|}\left(\cos\theta\right)\sin\left|m\right|\phi, & m<0,\\ \bar{P}_{\ell}^{0}\left(\cos\theta\right)/\sqrt{2}, & m=0,\\ \bar{P}_{\ell}^{m}\left(\cos\theta\right)\cos m\phi, & m>0. \end{cases}\]

The precision of the result may be increased by using arbitrary-precision arguments.

Examples

julia> Y = computeYlm(pi/2, 0, lmax = 1)
4-element SHArray(::Array{Complex{Float64},1}, (ML(0:1, -1:1),)):
     0.2820947917738782 + 0.0im
     0.3454941494713355 - 0.0im
 2.9918275112863375e-17 + 0.0im
    -0.3454941494713355 - 0.0im

julia> Y[(1,-1)] # index using (l,m)
0.3454941494713355 - 0.0im

julia> Y = computeYlm(big(pi)/2, big(0), lmax = big(1)) # Arbitrary precision
4-element SHArray(::Array{Complex{BigFloat},1}, (ML(0:1, -1:1),)):
    0.2820947917738781434740397257803862929220253146644994284220428608553212342207478 + 0.0im
    0.3454941494713354792652446460318896831393773703262433134867073548945156550201567 - 0.0im
 2.679783085063171668225419916118067917387251852939708540164955895366691604430101e-78 + 0.0im
   -0.3454941494713354792652446460318896831393773703262433134867073548945156550201567 - 0.0im

julia> computeYlm(SphericalHarmonics.NorthPole(), 0, lmax = 1)
4-element SHArray(::Array{Complex{Float64},1}, (ML(0:1, -1:1),)):
 0.2820947917738782 + 0.0im
               -0.0 + 0.0im
   0.48860251190292 + 0.0im
                0.0 + 0.0im

julia> Y = computeYlm(pi/2, pi/3, lmax = 1, m_range = SphericalHarmonics.ZeroTo, SHType = SphericalHarmonics.RealHarmonics())
3-element SHArray(::Array{Float64,1}, (ML(0:1, 0:1),)):
  0.2820947917738782
  2.9918275112863375e-17
 -0.24430125595146002

SphericalHarmonics.computeYlm! — Function

computeYlm!(Y::AbstractVector, θ::Real, ϕ::Real; lmax::Integer, [m::Integer] [m_range = SphericalHarmonics.FullRange], [SHType = SphericalHarmonics.ComplexHarmonics()])

Compute an entire set of spherical harmonics $Y_{l,m}(θ,ϕ)$ for $0 ≤ l ≤ l_\mathrm{max}$, and store them in the array Y.

The optional argument m_range decides if the spherical harmonics for negative m values are computed. By default the functions for all values of m are evaluated. Setting m_range to SphericalHarmonics.ZeroTo would result in only functions for m ≥ 0 being evaluated. Providing $m$ would override this, in which case only the polynomials corresponding to the azimuthal order $m$ would be computed.

The optional argument SHType may be used to choose between real and complex harmonics. To compute real spherical harmonics, set this to SphericalHarmonics.RealHarmonics().

SphericalHarmonics.computeYlm! — Function

computeYlm!(Y::AbstractVector, P::AbstractVector{<:Real}, θ::Real, ϕ::Real; lmax::Integer, [m::Integer], [m_range = SphericalHarmonics.FullRange], [SHType = SphericalHarmonics.ComplexHarmonics()])
computeYlm!(Y::AbstractVector, P::AbstractVector{<:Real}, θ::Real, ϕ::Real, lmax::Integer, [m::Integer, [m_range = SphericalHarmonics.FullRange, [SHType = SphericalHarmonics.ComplexHarmonics()]]])

Compute an entire set of spherical harmonics $Y_{l,m}(θ,ϕ)$ using the precomputed associated Legendre Polynomials $\bar{P}_l^m(\cos θ)$, and store in the array Y. The array P may be computed using computePlmcostheta.

The optional argument m_range decides if the spherical harmonics for negative m values are computed. By default the functions for all values of m are evaluated. Setting m_range to SphericalHarmonics.ZeroTo would result in only functions for m ≥ 0 being evaluated. Providing $m$ would override this, in which case only the polynomials corresponding to the azimuthal order $m$ would be computed.

The optional argument SHType may be used to choose between real and complex harmonics. To compute real spherical harmonics, set this to RealHarmonics().

SphericalHarmonics.compute_coefficients — Method

SphericalHarmonics.compute_coefficients(L::Integer)

Precompute coefficients $a_l^m$ and $b_l^m$ for all $2 ≤ l ≤ L$ and $0 ≤ m ≤ l-2$.

SphericalHarmonics.compute_coefficients(L::Integer, m::Integer)

Precompute coefficients $a_l^m$ and $b_l^m$ for all $|m| + 2 ≤ l ≤ L$ and the specified $m$.